Changing parameter order for 'trigger-builds' builder.
It seems that order is important for the various parameter options that
can be passed to the envinject parameterized trigger plugin.

The relative order (of XML generation) of the 'current-parameters'
option has been changed so that it is generated before any
'predefined-parameters'.  This allows for overriding any
of the current parameters with a pre-defined parameter.

Change-Id: If0429a63267b314c23e0ef331546e3187b4373c3
Reviewed-on: https://review.openstack.org/15050
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: Monty Taylor <mordred@inaugust.com>
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Tested-by: Jenkins