Add option to specify Zuul URL.
This is needed in order to run Zuul on a server other than
the Jenkins master.

This isn't a great design, but it is minimally invasive to JJB
while we work on improving Zuul to remove the need for parameter
specification and notifications.

Change-Id: Ib4dabfc3fe37623031bdbe4a8ec53f25a6070a8e
Reviewed-on: https://review.openstack.org/17517
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins