Adding extended-choice functionality
This change adds the ability to use the extended choice plugin.
This plugin allows you to populate a file with choices and then
allow the user to select one or more when the build now button
is clicked.  This change only changes the properties and setup
files.

Change-Id: I05af63f75f9ab87cd923c23177d33898c26149cf
Reviewed-on: https://review.openstack.org/20610
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Khai Do <zaro0508@gmail.com>
Approved: Monty Taylor <mordred@inaugust.com>
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Tested-by: Jenkins