Build when a change is pushed to GitHub
Added ability to build when a change is pushed to github, to the triggers.
This change is changing the triggers.py file and follows the poll scm
as a template for the change.  Setup.py was also changed to enable the new
feautre.  There is outside changes needed to Jenkins and/or github to fully
use this feature.  You need to tell Jenkins whether to automatically manage
the github repos or you need to put the correct URL into the webhook.
Furthermore, at the current time it is necessary to set both github and
pollscm for this to work.  In the future this may not be necessary so setting
github does not automatically set pollscm.

Change-Id: Id7b48fa47f63669ba31282a403f07a1234e7ac31
Reviewed-on: https://review.openstack.org/23433
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins