only generate XML for requested jobs
The updated and test commands generated the XML for all the jobs, even
if we were only interested by a specific job.  This patch pass the job
names requested to the job builder so the XML generation function can
skip jobs we are not interested in.  Note that we still have to expand
job templates to find out if the expanded job name match the filter.

Change-Id: I3db6bc3b07b0bfe71019c19c2d3d02bced1e25ac
Reviewed-on: https://review.openstack.org/26839
Reviewed-by: Jesse Keating <jesse.keating@rackspace.com>
Reviewed-by: Spencer Krum <krum.spencer@gmail.com>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins