Fixing ant java opts
The java-opts were being passed to ant with a new line in between
each option.  This causes ant to choke and throw a usage of java
message and then exit with a success.  This is not good.  Now the
java options are passed with a space in between each one.

Change-Id: I6921dfd396d91eed112442503afc1d35e942af6e
Reviewed-on: https://review.openstack.org/29127
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Khai Do <zaro0508@gmail.com>
Approved: James E. Blair <corvus@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins