job filtering was not working properly
When passing job names as arguments to 'update', the command is supposed
to only retain this jobs.  Due to the job being a dict, the filter would
never match and the none of the job would be updated.

This has apparently always been broken since the feature got introduced
in 85cf7a41.  Using job.['name'] fix it up.

Change-Id: Icf4d5b0bb68777f7faff91ade04451d4c8501c6a
Reviewed-on: https://review.openstack.org/34197
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: James E. Blair <corvus@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins