Use a more consistent style for booleans values.
For an optional boolean option, a common idiom was to write:
  data.get('option-name', 'false')

Using a keyword False instead of the string 'false' (resp. True
instead of 'true') is deemed clearer, since it conveys that the
option is a flag.

Such calls are modified to using False/True. The change has been
creating using `git grep '.\(false\|true\).)'` to find what needed
to change.

Change-Id: I54087b5487c7186fcc4ee70ea9b28c1a1002a547