Make sure tox does not call pip install with --pre
* tox.ini: By default tox passes --pre to pip install commands so
that prerelease packages will be deemed automatically suitable, as a
means of keeping backwards-compatible with earlier pip releases
which did not interpret version strings looking for potential
prerelease patterns. By overriding the install_command in tox 1.6
and later, this option can be removed from the pip command line so
that prerelease versions of packages will only be installed if a
version spec for a particular package is also a prerelease.

Change-Id: I978b906b0ecd6c01c99e08f9d27271a1f31c7a14