Implements: Archive publisher allow-empty setting.
The archive publisher did not allow setting the allowEmptyArchive setting in the YAML.
This change adds configuring the setting if it is defined in the YAML file. If the value
is 'true' then it is set, else it is 'false'. The default Jenkins behavior is 'false' disallowing
builds to pass without storing the archived data.

Updated the doc string to add that there is a default value (false) which is also the default Jenkins behavior.
Cleand up the value assignment, and typed it to a str().

fixed pep8 violation

Addresses patchset 3 request to update the doc string to use the test directly.
Would like to have more discussion around the second point in patchset 3 around having
the new code add a line of XML.

Change-Id: I5e0c9ed7e8ea669a8d0e8f267ebb02d5f0b6ae73