added logstash plugin support and fixed test typo
Added new test cases in wrappers fixtures.
Updated setup.py to include the new wrapper entry point.
Added the logstash plugin to the wrapper module.

made the default the localhost for host.
Added an if around the use redis bool.
If Redis parameters are not defined, use an empty dict to use all defaults.

Change-Id: Iee950a7d05a56d3a45b5f93ccd77306b37662d25