Install the jenkins_jobs package on setup
Since the conversion to PBR it looks like there is an omission of
the [files] section in the setup.cfg. This means that the metadata
was installed for the jenkins-job-builder but the package with the
python code was not. This patchset adds the [files] section in to
correct this oversight.

Closes-Bug: 1292745
Change-Id: Ia1b8f022a66574ee371587d49bb46c43d80f01ba