Treat ignore_cache as a boolean
When ignore_cache was loaded from the config file (including the
default config) it was treated as a string, which evaluates to
true in boolean expressions.  Therefore, unless ignore_cache was
set on the command line, it would always behave as if true.

Fix by asking configparser for a boolean value.

Change-Id: I63af8637671248fc1a6895d1eb82f1c2e24ce5c2