Allow `test` command to use conf files
It should be possible to run the `test` command with a custom config.
This allows examination of the XML outputted where config settings can
change the outputted XML. Otherwise any non-default config values will
only be used when running the `update` comand.

Add a simple test to confirm that the custom url that is unlikely to
ever be a valid default value is returned when the command is set to
`test` and a custom config is specified.

Change-Id: I193e83314a35943406bff5b182f20693edd640ba