Replace 'Testing JJB' with the term 'Test Mode'
The term 'testing' is overloaded.  I thought it meant running the actual
JJB tests instead of running the JJB application.  I think a good way to
distinguish the two is to use the term 'test mode' when it's in the
context of running the JJB application with the 'test' parameter.
I've already been using the term 'run in test mode' quite a bit in
my JJB reviews to distiguish it from the actual JJB tests.

Change-Id: I2076189e3ae55caa20d0894fb20859a7c461e46a