Get verison info for jenkins-job-builder
Old setuptools/distribute VersionInfo() needs the dist name not the
package name to get the correct version info. Switch to using
'jenkins-job-builder' which is the dist name instead of using
'jenkins_jobs' which is the package name so that we get the right data
when processing versions on old setuptools.

Closes-Bug: #1384238
Change-Id: I94a6bc6a5580ea8861eb603e5dfab2825475b45f