Document node parameter usage with matrix projects
The Matrix Tie Parent plugin has been deprecated for a while and the
documentation for it in JJB mentions that it's deprecated, but the
documentation doesn't call out how to use the new way of doing it.  This
patch updates the documentation to explicitly call out that the node
parameter only controls the matrix job's parent job.

Change-Id: Ie784f6d90667a0bc66ae0efd3eda1974ddbac4f2