Add details to credentials param in scm docs
The credentials-id parameter for git scm seem to be void of any details
or specifics like what exactly this parameter is, or where to get this
in jenkins. i was of opinion that its name of credential but took me a
while to fig out how to get credentials ID of a credential in jenkins.
Its actually last field of URL path when you click your credential in
 jenkins dashboard.
I have added a step-by-step github gist here:
https://gist.github.com/hayderimran7/78421229af85ae90177b
Also fixed repition of word 'create'.

Change-Id: I827779eeaae8b28f53764b70de6aaf4b80157547