zuul: drop notifications
Before Zuul migrated to Gearman, the jobs had to notify Zuul on start
and completion. That was done using the Jenkins notification plugin and
the configuration is done in JJB by using `triggers: [ zuul ]` which
also adds the build parameters to the job.

Since Zuul migrated to Gearman, the jenkins_endpoint listener (
http://127.0.0.1:8001/jenkins_endpoint ) has been removed so those HTTPs
requests are noops. Migration to Gearman happend in Zuul ver 2.0

Remove notifications entirely.
Keep the build parameters to retain the ability to manually trigger a build.
Adjust documentation accordingly.

Change-Id: Ifba876d001da90301b1efbfea3c2b8b031b42303