Remove superfluous calls to matches()
Remove unnecessary calls to matches() and ensure that delete_job() is
correctly passing the jobs_glob argument to the expandYaml() method call
since it uses a string whereas update_job() is using a sequence of
strings.

Once the bug with how delete_job is calling expandYaml() is fixed, the
subsequent call to matches() can be removed.

Change-Id: Iaf1404e26691384fd1d55ea630457860cc1beff5