Test with LANG set to UTF-8
Tox 2.0 does not pass the environment from the caller through anymore.
THis means your locale LANG setting is not set which makes the utf8
tests unhappy in JJB. Correct this by explicitly setting a LANG of
en_US.UTF-8. The important bit is the UTF-8 specification, en_US was
chosen somewhat arbitrarily.

Change-Id: I0fc5b5ab80414b392b7cebe9b4e0ebc5cc42c61c