Add support to reverse trigger for job list
Listing jobs as a comma separated list is fine with small amounts of
jobs however when this list gets too long it can be difficult to
maintain.

This patch adds support to list jobs as a list and the code will
automatically change it to a python separated list. Also retains support
for the old style comma separated list.

Change-Id: I0a419bb1f2ec83979f46990430f874381680cac8
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>