scm svn: add support for viewvc-url
This allows one to specify a URL that links to the web interface for the
source being built.

You can find the original version of this patch here:

  https://github.com/h01ger/jenkins-job-builder/commit/e38c41ef6e116072d9965f13cbb9d55eebe51f57

Update to the original author's patch is limited to fixing a very minor
pep8 errors, and adding the feature to a unit test.

It is being submitted upstream with Holger's approval, because he doesn't
have the time available to do it himself.

Change-Id: Ia5acadc5b7bb1a144cc375a692cb7ad0474a931f
Co-Authored-By: Philip Hands <phil@hands.com>