Adding support for Sidebar Plugin
<https://wiki.jenkins-ci.org/display/JENKINS/Sidebar+Plugin

You can find the original version of this patch here:
  https://github.com/h01ger/jenkins-job-builder/commit/c5ea266856d2d9205c4904e98dd3948269043a4e

Update to the original author's patch is limited to moving the one line
from setup.py to setup.cfg, fixing some pep8 errors, and adding a unit test.

It is being submitted upstream with Holger's approval, because he doesn't
have the time available to do it himself.

Change-Id: I75ffedeaafab3aea4eb1b172a980d625669fdacc
Co-Authored-By: Philip Hands <phil@hands.com>