fix default handling of script-file-path
Fixes the default handling of script-file-path property in
script trigger to more closely match the default behavior
when saving a job configuration in the Jenkins UI.

The previous behavior of setting script-file-path to an empty
string resulted in exceptions in the Jenkins log.

org.jenkinsci.plugins.scripttrigger.ScriptTriggerException:
The script file path '' doesn't exist.

Change-Id: I3d07892685c96824041efb102b2a474c22a2b9a7