Made a common helper function - convert_mapping_to_xml
This will allow other modules to convert a mapping set to xml easily
with defaults so other methods don't have to create their own.

Includes behaviour to selectively omit or include empty tags.

Behaviour for default value:
  None - omit this tag if users didn't put anything
  "" - create an empty element if users didn't put anything

This is applied to the existing openshift plugin functions in
refactoring as Jenkins actually retains an empty XML tag when the
optional 'auth-token' setting is missing from the input job definition.
Which results in modifying the default test's output for the openshift
plugin.

Change-Id: I935a70ad191bc0f3f3dbd571b2b1cf7468106697