Debug message leads to (n*m) log growth.
Where n = number of variables in scope for a particular job template instance.
and   m = number of job template instance.

Let's imagine for a large set of jobs with n = 100 and m = 50

This debug message makes it difficult to scan through JJB output looking for
signficant debug messages and does not provide a clear value.

It is trivially easy to know which variables are and which are not included in a
job-template's name by simply reading it.

Change-Id: I745e26ac6062277fc477cb1ca70cf2809f5806e3