Fix pollscm backwards compatiblity
This change https://review.openstack.org/#/c/211993/ made '' an
invalid attribute value for pollscm, but that is a valid value.  We use
'' for pollscm with the option to build when a changed is pushed to
github.  This might have changed since we set it up, but you had to
specify a blank poll schedule for jenkins to build on pushes to github.
This changes allows '' as a valid value.

Change-Id: I35c45b4619584b3358562131399fc5ef16c4c5ec