Add comment about proxy env pass through
For users in a proxy environment, checking the doc urls requires being
able to pass through the proxy environment variables. Include a simple
comment along with the tox command requiring it to help users checking
any links added are valid.

Add some references to documentation, both in the detailed
documentation and the quick README.rst that would be more commonly
checked.

Change-Id: I2316bc1b7e845c29a8b248eef4668a6099d109f7