project matrix: rename variable/refactor
While reviewing the YAML axis execution strategy for the matrix project,
Darragh Bailey suggested to:

* use proper variable naming
* JJB exceptions
* Rely on hudson_model
* Drop trailing slashes

Slightly refactor the matrix project method.

For the combination touchstones, the web interface list the result has
'Stable', 'Unstable'.  Map 'Stable' to 'Success' which is actually in
hudson.model.Result.  Add a basic test to cover that.

Change-Id: I7ab416a2dbadae6d8a41fe8902ebec04fc530abc