Fix documentation of --delete-old: affects only managed jobs.
As far back as its introduction (2cefa400), the `--delete-old` option
has only ever deleted jobs that were marked with the special
"<!-- Managed by Jenkins Job Builder -->" comment that JJB adds to
descriptions.

This change fixes the documentation of the `--delete-old` option,
which erroneously used to state that even jobs that were never
managed by JJB would be subject to deletion.

The test for the feature was slightly enriched as well: the fact
that unmanaged jobs were not deleted, was not actually tested.

Change-Id: I438a7b555b6e122869988b3a2b9ea40896004122