Change default on fail_required parameter to True for convert xml
If optional setting was set and then is removed, Jenkins will retain the last
configured setting for the option. This is inconsistent as it is expected that
Jenkins will revert back to its default setting. And so optional parameters
should use default setting. That is, JJB will handle default setting if
optional parameter is not set.

Original issue:
http://lists.openstack.org/pipermail/openstack-infra/2016-February/003746.html

Change-Id: Iaf8684053f8a2a035467f13d91fe2f0a56be5906
Signed-off-by: Kien Ha <kienha9922@gmail.com>