adding extra properties to sonar publisher
Allows you to explicitly set which sonar configuration to use
for the sonar analysis when having more than one sonar installation.
In this case you are required to select which one to use,
as there is no default fallback in the Jenkins plugin itself.

Also added a similar option for configuring the maven installation
to use. This does have a default value, the version used in the
maven build in the same job. However, if for some reason you
need a different version of maven for sonar analysis you
have to be able to configure which one.

Adds parameter installation-name for the sonar configuration,
and maven-installation-name for the maven configuration.

Change-Id: I013ea80688efa17334448e6394d312ca6854cf07