Regression test for update command with --delete-old
Ensure that we explicitly check that there were no extra calls in
addition to the expected set of calls to the delete jobs API, when
populating keep_jobs with the list of jobs updated and checking that
the unmanaged job is ignored.

Correct the list of jobs returned by get_jobs call.

Prevent issue introduced by: https://review.openstack.org/352060

Change-Id: I5b7452d593707db7085e7d305893f05a77be82a1