Add tox to test-requirements.txt
If I understand correctly, tox is already installed on automated test
nodes; however, the intent here is to improve local development workflow
by ensuring that an appropriate version is installed for running tests
locally.

This requires at least tox 2.9.1, which is the version that was used in
a zuul run observed today.

Change-Id: I8052e5792e229b58a48e48bd705ed576acdc2884