Add default columns to list views
This uses the default "All" view columns as the columns for any views
without a columns attribute sets.  This means that view definitions
which don't define any columns will still end up producing a useful
view.

(Note that the "All" columns are not fetched from the Jenkins instance;
the columns that are the default in a fresh Jenkins install are
hard-coded.)

Change-Id: I0a21e024be383088c6238c3fc7beb708af8c2c46