Migrate hipchat conf to new plugin style conf
Remove the old default hipchat configuration style in favour of
using the new plugin style and provide a mechanism to copy the url
value from the jenkins section automatically as part of a simple
migration.

This will allow removal of querying the jenkins section from existing
module code and allow for better isolation.

Change-Id: I6889777904ebabc01c044abcd31c9d8a20c255c4