Fix git-parameter sortMode values
It is impossible now to chose one of the following values of sortMode:
[ASCENDING_SMART, DESCENDING_SMART] because default value in JJB was hardcoded w/o '_',
but jenkins expect value w/ '_'. So if we pass e.g. ACENDING SMART (w/o '_') we get
improperly configured job and sortMode in this way replaced with NONE.
If we pass w/ '_' we got an error on testing JJB definitions.

Please merge ASAP.

Change-Id: I9bc4ff471ddb7d648005832e055ba14d59822489