Add `enhanced-logging` for `trigger-remote`
As for now, JJB lacks the ability to control the `Enable enhanced
logging` option of Parametrized Remote Trigger plugin.
This option allows the console log of the remote job to be printed in
the console log of the calling job.

This patch adds the option `enhanced-logging` to the `trigger-remote`
builder to control this checkbox.
The default value is set to `False` to keep the current behaviour for those
who do not need the logs.

Change-Id: I3d109603b4eb9ce247e5e6c4588053e1c61ba6ce