Fix Postbuildscript Plugin Info Lookup
Between version 0.18.1 and 2.0, the longName of postbuildscript changed
from 'Jenkins Post-Build Script Plug-in' to 'Jenkins PostBuildScript
Plugin'. Because of this JJB>1.6.2 currently does not work with previous
usages of the postbuildscript macros, as the lookup never succeeds for
the plugin<2.0.

This change switches the 'get_plugin_info' call to query using the
shortName, which hasn't changed.

Change-Id: I51437a300993ed09c7419bd9fc70afd0e8629119
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>