Refactor duplicated code in xml_config.py
The view generation code was largely a copy/paste of the job generation
code.  This refactors that shared code in to a super-class that both
code paths can share.

(This has the side effect of improving the error message when an
incorrect project-type is specified; that improvement was previously
only applied to views.)

Change-Id: I1316162b885267c5e21793e59c2c3f7b95d5ce3b